Add response codes to all API endpoints.
Added proper response codes and cleaned up the code of the controllers for the /users and /courses endpoints.
Added a ResourceNotFoundException, which represents a 404 Not found response code. The exception is a RuntimeException, made to be thrown whenever a resource doesn't exist in the database. Whenever it gets thrown from a method, a response gets returned with 404.
Cleaned up UserService and CourseService.