Implement a LoginService class.
The login service can be used to check a user account's existence and to validate a user's password when they try to log in.
The password is hashed using BCrypt. After that, the hashed password of the user is retrieved from the DB and the hashed passwords are compared. If they match, the user is valid, otherwise they aren't.