Implement registration and login logic.
Whenever a user registers, the username is checked. If it already exists, the server returns a message for that. If it doesn't exist, the user account gets created with the given username and password.
For the login process, the given username is checked for existence, after which the password is compared to the hashed password in the database. If they match, the login is successful.