Add basic authentication. (#7)
* Remove json web token dependencies.

* Add user_accounts table.

The user_accounts table is represented by the UserAccount class, which is a JPA Entity. Each user account has a one-to-one relationship with a user, and the user account is the owning side of the relationship.

* Change code to reflect the new changes to the User class.

Updated the classes that used to set/get the fields that were removed from the User class.

* Implement a LoginService class.

The login service can be used to check a user account's existence and to validate a user's password when they try to log in.
The password is hashed using BCrypt. After that, the hashed password of the user is retrieved from the DB and the hashed passwords are compared. If they match, the user is valid, otherwise they aren't.

* Update .gitignore.

* Change endpoint argument to Map.

Changed the authenticateUser endpoint to take a Map from the request body, as this allows for the body to have different key-value pairs such as the username and password, without having to create a DTO.

* Implement registration and login logic.

Whenever a user registers, the username is checked. If it already exists, the server returns a message for that. If it doesn't exist, the user account gets created with the given username and password.
For the login process, the given username is checked for existence, after which the password is compared to the hashed password in the database. If they match, the login is successful.

* Create AuthenticationService class.

The AuthenticationService class combines the authentication and registration logic and is going to replace LoginService and RegisterService.

* Change to use only AuthenticationService everywhere.

Deleted the LoginService and RegisterService classes and replaced them with the AuthenticationService class.